a little code cleanup.

+++ dbforms/src/org/dbforms/config/ResultSetVector.java
@@ -114,8 +114,7 @@ public class ResultSetVector
 
                   if (tmpObj != null)
                   {
-                     objectRow[i] = tmpObj.getSubString((long) 1, 
-                                                        (int) tmpObj.length());
+                     objectRow[i] = tmpObj.getSubString(1,(int) tmpObj.length());
                   }
                   else
                   {
