+++ dbforms/src/org/dbforms/util/external/PrintfFormat.java
@@ -2907,7 +2907,7 @@ public class PrintfFormat {
                   break;
             }
          } else {
-            sx = Integer.toString((int) x, 16);
+            sx = Integer.toString( x, 16);
          }
 
          return printXFormat(sx);
@@ -3332,7 +3332,7 @@ public class PrintfFormat {
                   break;
             }
          } else {
-            sx = Integer.toString((int) x, 8);
+            sx = Integer.toString(x, 8);
          }
 
          return printOFormat(sx);
